requirements.txt
streamlit
openai
python-dotenv
dspy-ai
yahooquery
